[mod/#6] viewModelFactory 파일 분리
