+++ app/src/main/java/com/example/letssopt/presentation/purchase/PurchaseViewModel.kt
@@ -1,9 +1,7 @@
 package com.example.letssopt.presentation.purchase
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import com.example.letssopt.core.data.local.dao.StoredItemDao
 import com.example.letssopt.core.data.local.entity.StoredItemEntity
 import com.example.letssopt.core.data.model.purchase.PurchaseContent
 import com.example.letssopt.core.data.repository.api.PurchaseRepository
@@ -45,18 +43,5 @@ class PurchaseViewModel(
                 _effect.send(PurchaseContract.Effect.ShowToast("저장에 실패했습니다."))
             }
         }
-    }
-}
-
-class PurchaseViewModelFactory(private val dao: StoredItemDao) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(PurchaseViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return PurchaseViewModel(
-                purchaseRepository = PurchaseRepository.getInstance(),
-                storageRepository = StorageRepository.getInstance(dao)
-            ) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
     }
 }
