+++ app/src/main/java/com/example/letssopt/presentation/storage/StorageViewModel.kt
@@ -2,9 +2,7 @@ package com.example.letssopt.presentation.storage
 
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
-import com.example.letssopt.core.data.local.dao.StoredItemDao
 import com.example.letssopt.core.data.local.entity.StoredItemEntity
 import com.example.letssopt.core.data.repository.api.StorageRepository
 import kotlinx.coroutines.Dispatchers
@@ -42,12 +40,3 @@ class StorageViewModel(private val storageRepository: StorageRepository) : ViewM
     }
 }
 
-class StorageViewModelFactory(private val dao: StoredItemDao) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(StorageViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return StorageViewModel(StorageRepository.getInstance(dao)) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
